requirements.txt
# Default: CUDA 11.8 GPU build.
# Equivalent explicit command:
# python -m pip install paddlepaddle-gpu==3.3.0 -i https://www.paddlepaddle.org.cn/packages/stable/cu118/

--extra-index-url https://www.paddlepaddle.org.cn/packages/stable/cu118/

python-pptx
PyMuPDF
opencv-python
numpy
Pillow
tkinterdnd2
pyinstaller
scikit-image
paddleocr
paddlepaddle-gpu==3.3.0
requests>=2.32.3
urllib3<2.0
charset_normalizer<4
chardet<6
